\name{multiModel}
\alias{multiModel}
\title{Test multiple models of character transition at a single node}
\description{
  Compares up to five different models of character evolution: two models of no transition (Brownian motion and a single-
  optimum Ornstein-Uhlenbeck model); one model allowing transition at a single node in a whole-tree context; and two models 
  that split the tree into two subtrees and treat character evolution separately in the two trees (the 'censored' approach -- see
  discussion below).
}
\usage{
multiModel(phy, dat, node, models = c("whole.brown", "whole.ou1", "whole.ou2", "part.brown", "part.ou"))
}
\arguments{
  \item{phy}{ An \pkg{ape}-style tree }
  \item{dat}{ A vector of data, with names corresponding to tips in \code{phy} }
  \item{node}{ A single node, defined by the tips that are descendent from it }
  \item{models}{ The vector of models, where \dQuote{whole} indicates a model in which the tree is treated as a single entity, and 
                  \dQuote{part} indicates a model in which the tree is subdivided and treated as two entities }
}
\details{
  The 'censored' approach was presented in O'Meara et al. 2006
}
\value{
  A list with two items:
  \item{IC}{A matrix of information criterion statistics, generated by \code{informationCriterion}}
  \item{modelMatrix}{A matrix of model parameters for each of the whole tree, partial tree, and summed-partial-tree models}
}
\references{  
  Hipp, A.L. (2007) 
  Non-uniform processes of chromosome evolution in sedges (Carex: Cyperaceae). 
  \emph{Evolution} 61:2175--2194.

  O'Meara, B. C., C. Ané, M. J. Sanderson, and P. C. Wainwright (2006) 
  Testing for different rates of continuous trait evolution using likelihood. 
  \emph{Evolution} 60:922--933.
}
\author{Andrew L. Hipp \email{ahipp@mortonarb.org}}
\note{
  In \pkg{ouch} version 2.x, the \code{ouchtree} function rescales trees to depth = 1. As a consequence, the alpha and sigma parameters
  are not directly comparable between subtrees, as they were in previous versions of \pkg{ouch}.
}
\seealso{ \code{\link{carex}} for an example}