\name{print.hansenSummary}
\alias{print.hansenSummary}
\title{Print method for a hansenSummary object}
\description{
  \code{print.hansenSummary} is a print method for a \code{hansenSummary} object.
}
\usage{
print(hansenSummary)
}
\arguments{
  \item{hansenSummary}{
  Output from \code{summary.hansenBatch}.
  }
}
\details{
    This function provides an easy-to-read synopsis of the analysis summary. Called automatically when you simply type
    in the name of the summary object.
}
\author{Andrew Hipp <ahipp@mortonarb.org>}
\seealso{
\code{\link{carex}} for example
}
